nimble/codexes-factory/pyproject.toml
[project]
name = "codexes-factory"
version = "0.2.0"
description = "AI-Assisted Publishing Lifecycle Platform"
requires-python = ">=3.12"
dependencies = [
    "python-docx>=1.1.0",
    "litellm>=1.35.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "PyMuPDF>=1.24.0",
    "streamlit>=1.34.0",
    "stripe",
    "isbnlib",
    "streamlit-authenticator>=0.4.2",
    "python-barcode>=0.15.1",
    "requests>=2.32.4",
    "jsonschema>=4.24.0",
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
    "psutil>=7.1.0",
    "beautifulsoup4>=4.12.0",
    "plotly>=6.3.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
